Change UI & add icons
