+++ resilience-front2/src/app/add-item/accordion-form/accordion-form.component.css
@@ -1,0 +1,13 @@
+input, textarea {
+    background-color: white;
+    width: 100%;
+}
+
+.big-class {
+  display: flex;
+  flex-direction: column;
+}
+
+.big-class > * {
+    width: 100%;
+}
