+++ resilience-front2/src/app/add-item/accordion-form/accordion-form.component.html
@@ -1,28 +1,36 @@
-<div class="row">
+<div class="row justify-content-md-left">
     <div class="col-md-6">
       <div class="form-group">
         <form (ngSubmit)="onSubmit(f)" #f="ngForm">
-          <label for="url"></label>
-          <input type="text" id="url" name="url" class="form-control" placeholder="URL" [(ngModel)]="url" required [value]="url">
-          <label for="name"></label>
-          <input type="text" id="name" name="name" class="form-control" placeholder="Name" [(ngModel)]="name" required [value]="name">
-          <label for="description"></label>
-          <input type="text" id="description" name="description" class="form-control" placeholder="Description" [(ngModel)]="description" required
-            [value]="description">
           <br>
-          <button class="btn btn-primary btn-lg col-md-12" type="submit" [disabled]="f.invalid"> {{ createOrUpdate }}</button>
+          <div class="big-class">
+          <mat-form-field>
+              <label for="name"></label>
+            <input matInput type="text" id="name" name="name" class="form-control" placeholder="Name" [(ngModel)]="name" required [value]="name">
+          </mat-form-field>
+          <br>
+          <mat-form-field>
+              <label for="url"></label>
+            <input matInput type="text" id="url" name="url" class="form-control" placeholder="URL" [(ngModel)]="url" required [value]="url">
+          </mat-form-field>
+          <br>
+          <mat-form-field>
+              <label for="description"></label>
+            <textarea matInput type="text" id="description" name="description" class="form-control" placeholder="Description" [(ngModel)]="description" required
+              [value]="description" rows="10"></textarea>
+          </mat-form-field>
+          <br>
+          <div>
+            <button class="btn btn-primary btn-lg col-md-12" type="submit" [disabled]="f.invalid"><span class="create"> {{ createOrUpdate }} &nbsp;</span><mat-icon>update</mat-icon></button>
+          </div> 
+        </div> 
         </form>
       </div>
     </div>
     <div class="col-md-3">
       <div class="add-category">
-        <button class="btn btn-primary btn-lg col-md-12 category">
-          <a [routerLink]="id" (click)="addCategory()">
-            <span class="category-text"> Add Category</span>
-          </a>
-        </button>
+        <app-add-category></app-add-category>
         <br>
-        <app-add-category *ngFor="let category of categories; let i = index" [index]="i" [id]="category.id" (categoryValue)="onSelected($event)"></app-add-category>
         <br>
       </div>
     </div>
