+++ resilience-front2/src/app/add-item/accordion-form/add-category/add-category.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
+
+@Component({
+  selector: 'app-add-category',
+  templateUrl: './add-category.component.html',
+  styleUrls: ['./add-category.component.css']
+})
+export class AddCategoryComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
