--- resilience-front2/src/app/add-item/add-category/add-category.component.html
@@ -1,7 +1,0 @@
-<div>
-  <select #selectedTag (click)="selectChangeHandler(selectedTag.value)">
-    <option value="tag1">tag1</option>
-    <option value="tag2">tag2</option>
-    <option value="tag3">tag3</option>
-  </select>
-</div>
