+++ resilience-front2/src/app/add-item/add-item.component.html
@@ -1,39 +1,18 @@
 <div class="container">
-  <div class="row justify-content-md-left">
+  <div class="row ">
     <div class="col-md-12">
-      <button class="btn btn-primary btn-lg col-md-4 add-reference"(click)="addReference()">
+      <button class="btn btn-primary btn-lg col-md-12 add-reference"(click)="addReference()">
         <a [routerLink]="id">
           <span class="category-text"> Add new reference</span>
         </a>
       </button>
-      <button class="btn btn-danger btn-lg offset-md-4 col-md-4 delete-reference">
-          <a [routerLink]="id">
-            <span class="category-text"> Delete references</span>
-          </a>
-      </button>
       <br>
     </div>
-  </div>
+  
   <br>
-  <!-- <ngb-accordion #acc="ngbAccordion" [closeOthers]="true" [activeIds]="lastCreated" type="default" *ngFor="let ref of refs; let i = index" (panelChange)="toggleAccordion($event, acc)">
-    <ngb-panel>
-        
-        <ng-template *ngIf="ref.name; else newReference" ngbPanelTitle>
-          <span>{{ ref.name }} </span>
-         
-        </ng-template>
-        <ng-template #newReference ngbPanelTitle>
-          <span>New Reference </span>
-          <button (click)="preventOpening($event)"><app-delete-item [id]="i"></app-delete-item></button>
-        </ng-template>
-        <ng-template ngbPanelContent>
-          <app-accordion-form [index]="i" [id]="ref.id" (submit)="isSubmitted($event, i)" (panel)="onTogglePanel($event, i)" [references]="ref" [createOrUpdate]="createOrUpdate"></app-accordion-form>
-        </ng-template>
-    </ngb-panel> -->
-  <!-- </ngb-accordion> -->
-
+      <div class="col-md-12">
   <mat-accordion class="example-headers-align" >
-      <mat-expansion-panel #accordion *ngFor="let ref of refs; let i = index" [id]="ref.id" [expanded]="step === i" (opened)="setStep(i)" multi="true">
+      <mat-expansion-panel #accordion *ngFor="let ref of refs; let i = index" [id]="ref.id" [expanded]="step === ref.id" (opened)="setStep(ref.id)" multi="true">
         <mat-expansion-panel-header>
           <mat-panel-title>
             <div *ngIf="ref.name; else newRef">
@@ -44,22 +23,32 @@
             </ng-template>
           </mat-panel-title>
           <mat-panel-description>
-            <div *ngIf="ref.url; else newUrl">
-                {{ ref.url }}
+            <div *ngIf="ref.url; else newUrl" class="url-forward-parent">
+              <a href="{{ ref.url }}" target="_blank" rel="noopener noreferrer">
+                <p class="url-forward"> {{ ref.url }} &nbsp;</p>
+                <mat-icon>open_in_browser</mat-icon>
+              </a>
             </div>
            <ng-template #newUrl>
               <span>New URL</span>
             </ng-template>
-           <button mat-button color="danger" (click)="deleteReference(i)"> X </button>
+           <button mat-icon-button color="warn" (click)="deleteReference(i)">
+             <mat-icon>delete_forever</mat-icon> 
+            </button>
           </mat-panel-description>
         </mat-expansion-panel-header>
         <ng-template matExpansionPanelContent>
           <app-accordion-form [index]="i" [id]="ref.id" (submit)="isSubmitted($event, i)" (panel)="onTogglePanel($event, i)" [references]="ref" [createOrUpdate]="createOrUpdate"></app-accordion-form>
           <mat-action-row>
-              <button mat-button color="warn" (click)="prevStep()">Previous</button>
-              <button mat-button color="primary" (click)="nextStep()">Next</button>
+              <button mat-icon-button color="basic" (click)="prevStep()"> 
+                <mat-icon>keyboard_arrow_left</mat-icon>
+              </button>
+              <button mat-icon-button color="basic" (click)="nextStep()">
+                  <mat-icon>keyboard_arrow_right</mat-icon>
+              </button>
           </mat-action-row>
         </ng-template>
       </mat-expansion-panel>
     </mat-accordion>
+  </div>
 </div>
